### 2020-05-04 #40 @bpbond
Ten Years Reproducibility Challenge: Drivers of Evapotranspiration from Boreal Wildfires
**Original article:** https://onlinelibrary.wiley.com/doi/full/10.1111/j.1365-2486.2008.01776.x

**Scientific domain:** Earth Sciences/Ecology
**Programming language:** R
