### 2020-08-30 #40 @jhmatthes
I successfully reproduced the results in the author's code. I have just one small suggestion to the paper: it would be helpful to specify which "fit statistics" were reported in the paper that were reproduced in this particular analysis.
