### 2020-11-09 #40 @karthik
Sorry for the delay @bpbond but I'll proceed with accepting your paper now. Stay tuned for next steps.
