### 2020-11-24 #40 @bpbond
Thanks, all.
