### 2021-02-01 #40 @rougier
@bpbond Sorry for the long delay, I'll handle the publication. Can you convert your article to the ReScience template?
