### 2021-03-11 #40 @bpbond
My sincere apologies—a busy last month but I am back and determined to finish this!

@rougier I did this article in R Markdown, but am not a LaTeX user and wondering what you suggest for converting to the ReScience template. Looking at the [template repository](https://github.com/ReScience/template), from what I understand one option is to install all the needed LaTeX tools locally, clone the repo, and use that. A second would be Overleaf (entering `metadata.tex` manually). Any recommendation?

Thanks.
