### 2021-03-14 #40 @bpbond
OK, I have moved the article content into a new Overleaf project, based on the ReScience template: https://www.overleaf.com/read/...

Attaching the resulting PDF here:
[Bond_Lamberty_ReScience.pdf](https://github.com/ReScience/submissions/files/6136685/Bond_Lamberty_ReScience.pdf)

@rougier Does this look good? What's my next step? Thanks for your help.
