### 2021-03-15 #40 @rougier
Look goods to me! I'll make a local copy of your overleaf and edit it to add publication infromation. In the meantime, can you save your code at software heritage (https://www.softwareheritage.org/save-and-reference-research-software/). The process is straightforward). You should obtain as swid that is needed for publication.

**NOTE:** I edited your post above to remove the overleaf link (else, anybody can modify your files)
