### 2021-03-15 #40 @bpbond
OK, it's up on Software Heritage: https://archive.softwareheritage.org/browse/revision/b8fe7506c23641ff7bc43deed4573feaa6d03947/?origin_url=https://github.com/bpbond/ten-years-paper33.git&timestamp=2021-03-15T15:55:00Z

I think this means the SWHID is `swh:1:rev:b8fe7506c23641ff7bc43deed4573feaa6d03947`.
