### 2021-03-18 #40 @rougier
@bpbond Here is a sandboxed version of your article: https://sandbox.zenodo.org/record/748080. Please tell me if everything looks right.
