+++ article/bibliography.bib
@@ -1,0 +1,261 @@
+
+% The entry below contains non-ASCII chars that could not be converted
+% to a LaTeX equivalent.
+@ARTICLE{Bond-Lamberty2009-py,
+  title    = "Effects of fire on regional evapotranspiration in the central
+              Canadian boreal forest",
+  author   = "Bond-Lamberty, Ben and Peckham, Scott D and Gower, Stith T and
+              Ewers, Brent E",
+  abstract = "Changes in fire regimes are driving the carbon balance of much of
+              the North American boreal forest, but few studies have examined
+              fire-driven changes in evapotranspiration (ET) at a regional
+              scale. This study used a version of the Biome-BGC process model
+              with dynamic and competing vegetation types, and explicit spatial
+              representation of a large (106 km2) region, to simulate the
+              effects of wildfire on ET and its components from 1948 to 2005 by
+              comparing the fire dynamics of the 1948--1967 period with those
+              of 1968--2005. Simulated ET averaged, over the entire temporal
+              and spatial modeling domain, 323 mm yr−1; simulation results
+              indicated that changes in fire in recent decades decreased
+              regional ET by 1.4\% over the entire simulation, and by 3.9\% in
+              the last ten years (1996--2005). Conifers dominated the
+              transpiration (EC) flux (120 mm yr−1) but decreased by 18\%
+              relative to deciduous broadleaf trees in the last part of the
+              20th century, when increased fire resulted in increased soil
+              evaporation, lower canopy evaporation, lower EC and a younger and
+              more deciduous forest. Well- and poorly-drained areas had similar
+              rates of evaporation from the canopy and soil, but EC was twice
+              as high in the well-drained areas. Mosses comprised a significant
+              part of the evaporative flux to the atmosphere (22 mm yr−1).
+              Modeled annual ET was correlated with net primary production, but
+              not with temperature or precipitation; ET and its components were
+              consistent with previous field and modeling studies. Wildfire is
+              driving significant changes in hydrological processes by
+              affecting mean stand age, forest species and energy balance.
+              These changes, particularly in poorly-drained areas, may control
+              the future carbon balance of the boreal forest.",
+  journal  = "Glob. Chang. Biol.",
+  volume   =  15,
+  number   =  5,
+  pages    = "1242--1254",
+  year     =  2009,
+  keywords = "black spruce; boreal forest; disturbance; evapotranspiration;
+              modeling; stand age; wildfire"
+}
+
+
+@ARTICLE{Thornton2005-hw,
+  title   = "Archiving numerical models of biogeochemical dynamics",
+  author  = "Thornton, Peter E and Cook, Robert B and Braswell, Bobby H and
+             Law, Beverly E and Post, Wilfred M and Shugart, Herman H and
+             Rhyne, B Timothy and Hook, Leslie A",
+  journal = "EOS",
+  volume  =  86,
+  number  =  44,
+  pages   = "431--432",
+  month   =  nov,
+  year    =  2005
+}
+
+@ARTICLE{Bond-Lamberty2007-uq,
+  title    = "Fire as the dominant driver of central Canadian boreal forest
+              carbon balance",
+  author   = "Bond-Lamberty, Ben and Peckham, Scott D and Ahl, Douglas E and
+              Gower, Stith T",
+  abstract = "Changes in climate, atmospheric carbon dioxide concentration and
+              fire regimes have been occurring for decades in the global boreal
+              forest, with future climate change likely to increase fire
+              frequency--the primary disturbance agent in most boreal forests.
+              Previous attempts to assess quantitatively the effect of changing
+              environmental conditions on the net boreal forest carbon balance
+              have not taken into account the competition between different
+              vegetation types on a large scale. Here we use a process model
[…]
