+++ article/metadata.yaml
@@ -24,7 +24,7 @@ affiliations:
     address: "College Park, MD, USA"
 
 # List of keywords (adding the programming language might be a good idea)
-keywords: rescience c, ten years reproducibility challenge, R, hydrology, wildfire
+keywords: Ten Years Reproducibility Challenge, R, Hydrology, Wildfire
 
 # Code URL and DOI/SWH (url is mandatory for replication, doi after acceptance)
 # You can get a DOI for your code from Zenodo, or an SWH identifier from
@@ -33,7 +33,7 @@ keywords: rescience c, ten years reproducibility challenge, R, hydrology, wildfi
 code:
   - url: https://github.com/bpbond/ten-years-paper33
   - doi:
-  - swh:
+  - swh: swh:1:rev:b8fe7506c23641ff7bc43deed4573feaa6d03947
 
 # Data URL and DOI (optional if no data)
 data:
@@ -48,13 +48,16 @@ replication:
  - doi:  10.1111/j.1365-2486.2008.01776.x
 
 # Don't forget to surround abstract with double quotes
-abstract: "As part of the ReScience Ten Years Reproducibility Challenge, I sought to reproduce the
-    analyses in Bond-Lamberty et al. (2009). The R program file for this paper was available, 
-    although it was unpublished and undocumented, but a full output file from the original analysis
-    was available. The results were essentially reproduced, with one or two minor changes in 
-    e.g. the sixth decimal place of a few numbers. The original choice to use only
-    base R functions imposed limitations but also enhanced reproducibility, given that software's
-    historical emphasis on backward compatibility."
+abstract: "As part of the ReScience Ten Years Reproducibility
+    Challenge, I sought to reproduce the analyses in Bond-Lamberty et
+    al. (2009). The R program file for this paper was available,
+    although it was unpublished and undocumented, but a full output
+    file from the original analysis was available. The results were
+    essentially reproduced, with one or two minor changes in e.g. the
+    sixth decimal place of a few numbers. The original choice to use
+    only base R functions imposed limitations but also enhanced
+    reproducibility, given that software's historical emphasis on
+    backward compatibility."
 
 # Bibliography file (yours)
 bibliography: bibliography.bib
@@ -64,7 +67,7 @@ bibliography: bibliography.bib
 #  * Editorial
 #  * Letter
 #  * Replication
-type: Reproduction / Ten Years Reproducibility Challenge
+type: Reproduction 
 
 # Scientific domain of the article (e.g. Computational Neuroscience)
 #  (one domain only & try to be not overly specific)
@@ -79,14 +82,14 @@ language: R
 
 # For example, the URL of the GitHub issue where review actually occured
 review:
-  - url:
+  - url: https://github.com/ReScience/submissions/issues/40
 
 contributors:
-  - name:
-    orcid:
+  - name: Karthik Ram
+    orcid: 0000-0002-0233-1757
     role: editor
-  - name:
-    orcid:
+  - name: Jaclyn Hatala Matthes
+    orcid: 0000-0001-8999-8062
     role: reviewer
   - name:
     orcid:
@@ -94,20 +97,20 @@ contributors:
 
 # This information will be provided by the editor
 dates:
-  - received:
-  - accepted:
-  - published:
+  - received: 04 May 2020
+  - accepted: 09 November 2020
+  - published: 18 March 2021
 
 # This information will be provided by the editor
 article:
-  - number: # Article number will be automatically assigned during publication
-  - doi:    # DOI from Zenodo
-  - url:    # Final PDF URL (Zenodo or rescience website?)
+  - number: 21
+  - doi:    
+  - url:    
 
 # This information will be provided by the editor
[…]
