### 2021-03-18 #40 @rougier
It's online ! See https://rescience.github.io/bibliography/Bond-Lamberty_2021.html
